# node-ouput/swagger.yaml
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "The Social Media",
    "description" : "API Documentation for The Social Media App",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000/api/v1"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
    "/users/signup" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Register a new user",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/users_signup_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "User created successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/users/login" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Login a user",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/users_login_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "User logged in successfully"
          },
          "401" : {
            "description" : "Invalid credentials"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/users/get-user-data" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get logged in user data",
        "responses" : {
          "200" : {
            "description" : "User data fetched successfully"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users/user/all" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get all users",
        "responses" : {
          "200" : {
            "description" : "Users fetched successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get all users",
        "responses" : {
          "200" : {
            "description" : "Users fetched successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users/user/{id}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get user by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The user ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "User fetched successfully"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users/edit-profile" : {
      "put" : {
        "tags" : [ "Users" ],
        "summary" : "Edit user profile",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/users_editprofile_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Profile updated successfully"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users/follow/{id}" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Follow a user",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The user ID to follow",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully followed the user"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users/unfollow/{id}" : {
      "delete" : {
        "tags" : [ "Users" ],
        "summary" : "Unfollow a user",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The user ID to unfollow",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully unfollowed the user"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/users/followers/{id}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get user's followers",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The user ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Followers fetched successfully"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "users_signup_body" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "John Doe"
          },
          "email" : {
            "type" : "string",
            "example" : "john@example.com"
          },
          "username" : {
            "type" : "string",
            "example" : "johndoe"
          },
          "password" : {
            "type" : "string",
            "example" : "password123"
          }
        }
      },
      "users_login_body" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "example" : "johndoe"
          },
          "password" : {
            "type" : "string",
            "example" : "password123"
          }
        }
      },
      "users_editprofile_body" : {
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string",
            "example" : "John"
          },
          "lastName" : {
            "type" : "string",
            "example" : "Doe"
          },
          "username" : {
            "type" : "string",
            "example" : "johndoe"
          },
          "email" : {
            "type" : "string",
            "example" : "john@example.com"
          },
          "photo" : {
            "type" : "string",
            "format" : "binary"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}